api: tcmu_cdb_print_info convert err log to debug
plenty of WARN logs:

2019-06-12 17:48:46.779 8794 [WARN] tcmur_cmdproc_thread:629 : Command
0xa3 not supported
2019-06-12 17:48:46.779 8794 [WARN] tcmur_cmdproc_thread:629 : Command
0xa3 not supported
2019-06-12 17:48:46.782 8794 [WARN] tcmur_cmdproc_thread:629 : Command
0xa3 not supported
2019-06-12 17:48:46.782 8794 [WARN] tcmur_cmdproc_thread:629 : Command
0xa3 not supported
2019-06-12 17:48:46.782 8794 [WARN] tcmur_cmdproc_thread:629 : Command
0xa3 not supported
2019-06-12 17:48:46.787 8794 [WARN] tcmur_cmdproc_thread:629 : Command
0xa3 not supported

Yet another case, where we get a lot of logs like above, reporting
MAINTENANCE_IN command not supported.

And I believe we all know that we don't support it. And as Mike rightly
mentioned in the conversation at PR#563, the initiator does not care if these
commands fail because it is not being supported. And users don't have to be
aware of them.

Getting them when running under debug mode is cool.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>