Prep for separating runner fields from tcmulib_cmd
The next patch moves the runner specific fields from tcmulib_cmd to a
new tcmur_cmd struct. The tcmu-runner callouts will be taking the new
struct, so this patch just performs the renaming part that is not
dependant on libtcmu_cmd, so the next patch is a little smaller and
easier to read.

Signed-off-by: Mike Christie <mchristi@redhat.com>