Add tcmur-runner cmd struct
This adds the tcmu-runner cmd struct which will be used by
tcmur_cmd_handler.c to hold fields specific to tcmu-runner. It also adds
some helpers that will be used to setup the cmd specific state and
manage their iovecs.

Note: This patch is not git bisectable. I separated it out to make it
easier to read.

Signed-off-by: Mike Christie <mchristi@redhat.com>