rbd: fix logging when client is blacklisted
If the client is blacklisted then don't log an error messages. This is
expected behavior if the initiator has failed over to us, then failed
back, and then the initiator sends test/probe commands like RTPG,
INQUIRY, etc. The target/device layer will handle this case by
initiating a reopen if/when needed or report the correct behavior for
the command that is being executed.

Signed-off-by: Mike Christie <mchristi@redhat.com>