runner: fix thread leak
We make the tpg on demand during recovery and free it when we are done.
We then never reap the recovery thread, so we could end up running out
of resources. This just has us use the detached state.

Note:
Because this is not device based we do not use the tcmur_work helpers.
The tpg is just allocated when we create the thread then delete it from
the thread, so we do not have any context to delete the work if we are
to use it.

Signed-off-by: Mike Christie <mchristi@redhat.com>