runner: fix alua state check during runner restart
If tcmu-runner restarts before the initiator side path failure timer
fires then the initiator will use the path without revalidating it. This
adds a check when the first IO is run and forces us to update the alua
and lock state if we do not get a INQUIRY or RTPG before IO restarts.

Signed-off-by: Mike Christie <mchristi@redhat.com>