tcmur_device: fix racy between reopening devices and reporting events
To make sure all the in-flight IOs have been finished before flushing
the event_work. Or just after we flush the event_work a new timedout
IO callback could come and then it will fire a new event work, which
may access the device while we are closing the device later.

Signed-off-by: Xiubo Li <xiubli@redhat.com>