tcmur_device: skip reporting events if the device is closed
After the rbd device being blocklisted it will lost the connection
and at the same time will timedout all the IOs. While the tcmu-runner
will report a 'lock_lost_cnt' event to ceph cluster. And if during
this the device is reopened it will cause the use-after-free bug.

When reopening the device it will hold the 'rdev->rdev_lock' lock and
then clear the 'TCMUR_DEV_FLAG_IS_OPEN' flag. And we can check this
flag and just skip reporting events if the device is closed.

Signed-off-by: Xiubo Li <xiubli@redhat.com>